services/dns_ssl_service/requirements.txt
flask>=2.3.0
dnspython>=2.4.0
prometheus-client>=0.20.0